backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Community data scrapers (Douban, WeRead, Flomo)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "uvicorn>=0.44.0",
    "playwright>=1.58.0",
    "pydantic[email]>=2.13.2",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "websockets>=16.0",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "pyjwt>=2.8",
    "bcrypt>=4.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/community", "db"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
